chore: move view out of _internal
